requirements.txt
alembic==1.4.3
autopep8==1.5.4
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
lazy-object-proxy==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.11.1
python-dotenv==0.14.0
Authlib==0.15.2